wip: break up into pieces
